nutrilift/backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
PyJWT==2.10.1
django-cors-headers==4.6.0
python-decouple==3.8
psycopg2-binary==2.9.10
hypothesis==6.92.1
requests==2.32.3
